+++ src/main/java/com/tobiasdiez/easybind/monadic/MonadicBinding.java
@@ -1,6 +1,0 @@
-package com.tobiasdiez.easybind.monadic;
-
-import javafx.beans.binding.Binding;
-
-public interface MonadicBinding<T>
-extends Binding<T>, MonadicObservableValue<T> {}