+++ src/main/java/com/tobiasdiez/easybind/optional/OptionalBinding.java
@@ -1,0 +1,8 @@
+package com.tobiasdiez.easybind.optional;
+
+import java.util.Optional;
+
+import javafx.beans.binding.Binding;
+
+public interface OptionalBinding<T> extends Binding<Optional<T>>, ObservableOptionalValue<T> {
+}