+++ src/main/java/com/tobiasdiez/easybind/select/IntermediateSelectedBuilder.java
@@ -4,7 +4,7 @@ import java.util.function.Function;
 
 import javafx.beans.value.ObservableValue;
 
-import com.tobiasdiez.easybind.monadic.MonadicBinding;
+import com.tobiasdiez.easybind.EasyBinding;
 
 class IntermediateSelectedBuilder<T, U> implements ParentSelectedBuilder<U> {
     private final ParentSelectedBuilder<T> parent;
@@ -18,7 +18,7 @@ class IntermediateSelectedBuilder<T, U> implements ParentSelectedBuilder<U> {
     }
 
     @Override
-    public <V> MonadicBinding<V> create(
+    public <V> EasyBinding<V> create(
             NestedSelectionElementFactory<U, V> nestedSelectionFactory) {
         NestedSelectionElementFactory<T, V> intermediateSelectionFactory = onInvalidation -> {
             return new IntermediateSelectionElement<T, U, V>(onInvalidation, selector, nestedSelectionFactory);