+++ src/main/java/com/tobiasdiez/easybind/select/ParentSelectedBuilder.java
@@ -4,7 +4,7 @@ import java.util.function.Function;
 
 import javafx.beans.value.ObservableValue;
 
-import com.tobiasdiez.easybind.monadic.MonadicBinding;
+import com.tobiasdiez.easybind.EasyBinding;
 
 interface ParentSelectedBuilder<T> extends SelectBuilder<T> {
     @Override
@@ -13,12 +13,12 @@ interface ParentSelectedBuilder<T> extends SelectBuilder<T> {
     }
 
     @Override
-    default <U> MonadicBinding<U> selectObject(Function<? super T, ObservableValue<U>> selector) {
+    default <U> EasyBinding<U> selectObject(Function<? super T, ObservableValue<U>> selector) {
         NestedSelectionElementFactory<T, U> leafSelectionFactory = onInvalidation -> {
             return new LeafSelectionElement<T, U>(onInvalidation, selector);
         };
         return create(leafSelectionFactory);
     }
 
-    <U> MonadicBinding<U> create(NestedSelectionElementFactory<T, U> nestedSelectionFactory);
+    <U> EasyBinding<U> create(NestedSelectionElementFactory<T, U> nestedSelectionFactory);
 }