+++ src/main/java/com/tobiasdiez/easybind/select/RootSelectedBuilder.java
@@ -2,7 +2,7 @@ package com.tobiasdiez.easybind.select;
 
 import javafx.beans.value.ObservableValue;
 
-import com.tobiasdiez.easybind.monadic.MonadicBinding;
+import com.tobiasdiez.easybind.EasyBinding;
 
 class RootSelectedBuilder<T> implements ParentSelectedBuilder<T> {
     private final ObservableValue<T> root;
@@ -12,7 +12,7 @@ class RootSelectedBuilder<T> implements ParentSelectedBuilder<T> {
     }
 
     @Override
-    public <U> MonadicBinding<U> create(
+    public <U> EasyBinding<U> create(
             NestedSelectionElementFactory<T, U> nestedSelectionFactory) {
         return new SelectObjectBinding<T, U>(root, nestedSelectionFactory);
     }