+++ src/main/java/com/tobiasdiez/easybind/select/SelectBuilder.java
@@ -4,11 +4,12 @@ import java.util.function.Function;
 
 import javafx.beans.value.ObservableValue;
 
-import com.tobiasdiez.easybind.monadic.MonadicBinding;
+import com.tobiasdiez.easybind.EasyBinding;
 
 public interface SelectBuilder<T> {
     <U> SelectBuilder<U> select(Function<? super T, ObservableValue<U>> selector);
-    <U> MonadicBinding<U> selectObject(Function<? super T, ObservableValue<U>> selector);
+
+    <U> EasyBinding<U> selectObject(Function<? super T, ObservableValue<U>> selector);
 
     static <T> SelectBuilder<T> startAt(ObservableValue<T> selectionRoot) {
         return new RootSelectedBuilder<T>(selectionRoot);