+++ src/main/java/module-info.java
@@ -1,6 +1,6 @@
 module com.tobiasdiez.easybind {
     exports com.tobiasdiez.easybind;
-    exports com.tobiasdiez.easybind.monadic;
+    exports com.tobiasdiez.easybind.optional;
     exports com.tobiasdiez.easybind.select;
 
     requires javafx.base;